Evolve MinimalStreamerSoA

+++ DataFormats/MinimalStreamer/interface/MinimalStreamerSoA.h
@@ -3,8 +3,8 @@
 #include "DataFormats/SoATemplate/interface/SoALayout.h"
 
 GENERATE_SOA_LAYOUT(MinimalStreamerSoATemplate,
-                    SOA_COLUMN(float, x),
-                    SOA_COLUMN(float, y),
-                    SOA_COLUMN(float, z))
+                    SOA_COLUMN(double, x),
+                    SOA_COLUMN(double, y),
+                    SOA_COLUMN(double, z))
 
 using SoA = MinimalStreamerSoATemplate<>;
